requirements.txt
arrow==0.10.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-forms-bootstrap==3.1.0
django-paypal==0.3.6
idna==2.6
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
stripe==1.62.1
urllib3==1.22
git+https://github.com/robinzigmond/reusable_blog_app.git
